+++ puffin_viewer/src/main.rs
@@ -126,7 +126,10 @@ fn main() {
         }
     };
 
-    let options = Default::default();
+    let options = epi::NativeOptions {
+        drag_and_drop_support: true,
+        ..Default::default()
+    };
     eframe::run_native(Box::new(app), options);
 }