+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## Unreleased
+* Implement `Hash` on `ThreadInfo`.
 
 
 ## 0.9.0 - 2021-09-20